+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [Unreleased]
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+- **PODAAC-2552**
+    - Catch exceptions during 'PerformFunction' and always send a failure response message, before re-throwing the 
+      exception.
+### Security
+    
 # [v1.2.0] - 2020-12-15
 ### Added
 - **PODAAC-2783**